+++ src/components/Button/CHANGELOG.md
@@ -20,3 +20,4 @@
 - 2025-11-23: Update the CSS variables names with camelCase.
 - 2025-11-26: Add `link` variant.
 - 2025-12-13: Update to support the new color scheme.
+- 2026-02-18: Replace Button stylesheet module from `scss` to `css`.